backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.10.0
pydantic-ai==0.0.15
python-dotenv==1.0.0
httpx>=0.27.2
openai>=1.40.6
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
PyJWT==2.8.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
email-validator==2.1.0
requests>=2.31.0
redis>=5.0.0
upstash-redis>=0.15.0
langchain>=0.1.0
langchain-openai>=0.0.6
langchain-community>=0.0.20
pyotp>=2.9.0
python-jose[cryptography]>=3.3.0
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
fpdf2>=2.7.0
supabase>=2.0.0
pgvector>=0.2.5
tiktoken>=0.5.0
pypdf>=3.17.0